Fixed circular imports between names.ts and stats.ts, and cleaned up display stats filtering to eliminate unnecessary creation of new UnitStat objects.
 On branch stats_refactor
 Changes to be committed:
	modified:   ui/core/components/character_stats.tsx
	modified:   ui/core/proto_utils/names.ts
	modified:   ui/core/proto_utils/stats.ts
