+++ ui/core/proto_utils/names.ts
@@ -1,7 +1,6 @@
 import { ResourceType } from '../proto/api.js';
 import { ArmorType, Class, ItemSlot, Profession, PseudoStat, Race, RangedWeaponType, Spec, Stat, WeaponType } from '../proto/common.js';
 import { DungeonDifficulty, RaidFilterOption, RepFaction, RepLevel, SourceFilterOption, StatCapType } from '../proto/ui.js';
-import { UnitStat } from './stats.js';
 
 export const armorTypeNames: Map<ArmorType, string> = new Map([
 	[ArmorType.ArmorTypeUnknown, 'Unknown'],
@@ -109,41 +108,6 @@ export function nameToProfession(name: string): Profession {
 	}
 	return Profession.ProfessionUnknown;
 }
-
-export const displayStatOrder: Array<UnitStat> = [
-	UnitStat.fromStat(Stat.StatHealth),
-	UnitStat.fromStat(Stat.StatMana),
-	UnitStat.fromStat(Stat.StatArmor),
-	UnitStat.fromStat(Stat.StatBonusArmor),
-	UnitStat.fromStat(Stat.StatStamina),
-	UnitStat.fromStat(Stat.StatStrength),
-	UnitStat.fromStat(Stat.StatAgility),
-	UnitStat.fromStat(Stat.StatIntellect),
-	UnitStat.fromStat(Stat.StatSpirit),
-	UnitStat.fromStat(Stat.StatSpellPower),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatSpellHitPercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatSpellCritPercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatSpellHastePercent),
-	UnitStat.fromStat(Stat.StatSpellPenetration),
-	UnitStat.fromStat(Stat.StatMP5),
-	UnitStat.fromStat(Stat.StatAttackPower),
-	UnitStat.fromStat(Stat.StatRangedAttackPower),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatPhysicalHitPercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatPhysicalCritPercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatMeleeHastePercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatRangedHastePercent),
-	UnitStat.fromStat(Stat.StatExpertiseRating),
-	UnitStat.fromStat(Stat.StatMasteryRating),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatBlockPercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatDodgePercent),
-	UnitStat.fromPseudoStat(PseudoStat.PseudoStatParryPercent),
-	UnitStat.fromStat(Stat.StatResilienceRating),
-	UnitStat.fromStat(Stat.StatArcaneResistance),
-	UnitStat.fromStat(Stat.StatFireResistance),
-	UnitStat.fromStat(Stat.StatFrostResistance),
-	UnitStat.fromStat(Stat.StatNatureResistance),
-	UnitStat.fromStat(Stat.StatShadowResistance),
-];
 
 export function getStatName(stat: Stat): string {
 	if (stat == Stat.StatRangedAttackPower) {
